move all authentication files into one folder and delete useless files
