+++ Cyclops/src/app/app.component.ts
@@ -1,7 +1,8 @@
 
 import { Component } from '@angular/core';
 import { Router } from '@angular/router';
-import { AuthService } from './auth/auth.service';
+import { AuthService } from './authentication/auth/auth.service';
+
 
 @Component({
   selector: 'app-root',
