+++ Cyclops/src/app/authentication/loginComponents/icon-button/icon-button.component.scss
