+++ Cyclops/src/app/authentication/registration/registration.module.ts
