+++ Cyclops/src/app/authentication/registration/registration.page.scss
