+++ Cyclops/src/app/authentication/reset-password/reset-password.page.scss
