--- Cyclops/src/app/crud-demo/crud-demo.page.html
@@ -1,31 +1,0 @@
-<ion-header>
-  <ion-toolbar>
-    <ion-title>CRUD Demo</ion-title>
-  </ion-toolbar>
-</ion-header>
-
-<ion-content>
-  <ion-list lines="full">
-    <ion-item>
-      <ion-label position="floating">Name</ion-label>
-      <ion-input [(ngModel)]="name" type="text"></ion-input>
-    </ion-item>
-    <ion-row>
-      <ion-col>
-        <ion-button type="button" (click)="addName()"> Add Name</ion-button>
-      </ion-col>
-    </ion-row>
-  </ion-list>
-
-  <ion-list>
-    <ion-list-header>Name List</ion-list-header>
-
-    <ion-item lines="inset" *ngFor="let user of users">
-      <ion-label>
-        <h4>{{user.id}} ---- {{user.name}}</h4>
-      </ion-label>
-    </ion-item>
-    
-  </ion-list>
-
-</ion-content>
