--- Cyclops/src/app/crud-demo/crud-demo.page.spec.ts
@@ -1,24 +1,0 @@
-import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
-import { IonicModule } from '@ionic/angular';
-
-import { CRUDDemoPage } from './crud-demo.page';
-
-describe('CRUDDemoPage', () => {
-  let component: CRUDDemoPage;
-  let fixture: ComponentFixture<CRUDDemoPage>;
-
-  beforeEach(waitForAsync(() => {
-    TestBed.configureTestingModule({
-      declarations: [ CRUDDemoPage ],
-      imports: [IonicModule.forRoot()]
-    }).compileComponents();
-
-    fixture = TestBed.createComponent(CRUDDemoPage);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  }));
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
