--- Cyclops/src/app/find-password/find-password-routing.module.ts
@@ -1,17 +1,0 @@
-import { NgModule } from '@angular/core';
-import { Routes, RouterModule } from '@angular/router';
-
-import { FindPasswordPage } from './find-password.page';
-
-const routes: Routes = [
-  {
-    path: '',
-    component: FindPasswordPage
-  }
-];
-
-@NgModule({
-  imports: [RouterModule.forChild(routes)],
-  exports: [RouterModule],
-})
-export class FindPasswordPageRoutingModule {}
