--- Cyclops/src/app/find-password/find-password.page.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-find-password',
-  templateUrl: './find-password.page.html',
-  styleUrls: ['./find-password.page.scss'],
-})
-export class FindPasswordPage implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
